+++ CHANGELOG.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.1.0 (unreleased)
+1.1.0 (2017-02-23)
 ------------------
 
 - Upgrade to last storage, permissions and cache backends features. (#7)